fix(audit-manager): move form to ui lib
