+++ apps/audit-manager/src/app/app.routes.ts
@@ -1,11 +1,6 @@
 import { Route } from '@angular/router';
-// eslint-disable-next-line @nx/enforce-module-boundaries
-import { auditBuilderRoutes } from '../../../../libs/feature/audit-builder/src/lib/audit-builder.routes';
+import { auditBuilderRoutes } from '@app-speed/feature/audit-builder';
 
 export const appRoutes: Route[] = [
-  {
-    path: 'audit-builder',
-    loadComponent: () => import('@app-speed/feature/audit-builder').then(i => i.AuditBuilderComponent)
-  },
   ...auditBuilderRoutes
 ];
