+++ libs/ui/audit-builder/src/audit-builder/audit-builder.component.scss
