+++ libs/ui/audit-builder/src/audit-form-step/audit-form-step.component.scss
