+++ libs/ui/audit-builder/src/audit-globals/audit-globals.component.scss
