+++ tsconfig.base.json
@@ -27,6 +27,7 @@
       "features/simple-audit": ["libs/features/simple-audit/src/index.ts"],
       "shared": ["libs/shared/src/index.ts"],
       "ui": ["libs/ui/src/index.ts"],
+      "ui/audit-builder": ["libs/ui/audit-builder/src/index.ts"],
       "ui/shell": ["libs/ui/shell/src/index.ts"]
     }
   }
